Extract SSL config properties from ThriftMetastoreConfig
