+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/metastore/thrift/TestThriftMetastoreConfig.java
@@ -47,10 +47,6 @@ public class TestThriftMetastoreConfig
                 .setMaxBackoffDelay(new Duration(1, SECONDS))
                 .setMaxRetryTime(new Duration(30, SECONDS))
                 .setTlsEnabled(false)
-                .setKeystorePath(null)
-                .setKeystorePassword(null)
-                .setTruststorePath(null)
-                .setTruststorePassword(null)
                 .setImpersonationEnabled(false)
                 .setUseSparkTableStatisticsFallback(true)
                 .setDelegationTokenCacheTtl(new Duration(1, HOURS))
@@ -79,10 +75,6 @@ public class TestThriftMetastoreConfig
                 .put("hive.metastore.thrift.client.max-backoff-delay", "4s")
                 .put("hive.metastore.thrift.client.max-retry-time", "60s")
                 .put("hive.metastore.thrift.client.ssl.enabled", "true")
-                .put("hive.metastore.thrift.client.ssl.key", keystoreFile.toString())
-                .put("hive.metastore.thrift.client.ssl.key-password", "keystore-password")
-                .put("hive.metastore.thrift.client.ssl.trust-certificate", truststoreFile.toString())
-                .put("hive.metastore.thrift.client.ssl.trust-certificate-password", "truststore-password")
                 .put("hive.metastore.thrift.impersonation.enabled", "true")
                 .put("hive.metastore.thrift.delegation-token.cache-ttl", "1d")
                 .put("hive.metastore.thrift.delegation-token.cache-maximum-size", "9999")
@@ -104,10 +96,6 @@ public class TestThriftMetastoreConfig
                 .setMaxBackoffDelay(new Duration(4, SECONDS))
                 .setMaxRetryTime(new Duration(60, SECONDS))
                 .setTlsEnabled(true)
-                .setKeystorePath(keystoreFile.toFile())
-                .setKeystorePassword("keystore-password")
-                .setTruststorePath(truststoreFile.toFile())
-                .setTruststorePassword("truststore-password")
                 .setImpersonationEnabled(true)
                 .setDelegationTokenCacheTtl(new Duration(1, DAYS))
                 .setDelegationTokenCacheMaximumSize(9999)
